migrated test_conserved_moments_d2q9
